# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "BundeslandResponse": {
        "properties": {
          "beziehung_start": {
            "title": "Beziehung Start",
            "type": "integer"
          },
          "bundesrat_fraktion": {
            "title": "Bundesrat Fraktion",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "koalition": {
            "items": {
              "type": "string"
            },
            "title": "Koalition",
            "type": "array"
          },
          "min_complexity": {
            "title": "Min Complexity",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "partei": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partei"
          },
          "stimmgewicht": {
            "title": "Stimmgewicht",
            "type": "integer"
          },
          "themen": {
            "items": {
              "type": "string"
            },
            "title": "Themen",
            "type": "array"
          },
          "wirtschaft_typ": {
            "title": "Wirtschaft Typ",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "koalition",
          "bundesrat_fraktion",
          "wirtschaft_typ",
          "themen",
          "beziehung_start",
          "stimmgewicht",
          "min_complexity"
        ],
        "title": "BundeslandResponse",
        "type": "object"
      },
...
      "BundesratResponse": {
        "properties": {
          "basis_bereitschaft": {
            "title": "Basis Bereitschaft",
            "type": "integer"
          },
          "beziehung_start": {
            "title": "Beziehung Start",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "laender": {
            "items": {
              "type": "string"
            },
            "title": "Laender",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sonderregel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sonderregel"
          },
          "sprecher_bio": {
            "title": "Sprecher Bio",
            "type": "string"
          },
          "sprecher_color": {
            "title": "Sprecher Color",
            "type": "string"
          },
          "sprecher_initials": {
            "title": "Sprecher Initials",
            "type": "string"
          },
          "sprecher_land": {
            "title": "Sprecher Land",
            "type": "string"
          },
          "sprecher_name": {
            "title": "Sprecher Name",
            "type": "string"
          },
          "sprecher_partei": {
            "title": "Sprecher Partei",
            "type": "string"
          },
          "tradeoffs": {
            "items": {
              "$ref": "#/components/schemas/BundesratTradeoffResponse"
            },
            "title": "Tradeoffs",
            "type": "array"
          }
        },
        "required": [
          "id",
          "laender",
          "basis_bereitschaft",
          "beziehung_start",
          "sprecher_initials",
          "sprecher_color",
          "name",
          "sprecher_name",
          "sprecher_partei",
          "sprecher_land",
          "sprecher_bio",
          "tradeoffs"
        ],
        "title": "BundesratResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MedienAkteurResponse": {
        "description": "SMA-392: Medienakteur-Definitionen (Seed aus DB), lokalisierter Name.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "min_complexity": {
            "default": 2,
            "title": "Min Complexity",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reichweite": {
            "title": "Reichweite",
            "type": "number"
          },
          "stimmung_start": {
            "default": 0,
            "title": "Stimmung Start",
            "type": "integer"
          },
          "typ": {
            "title": "Typ",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "typ",
          "reichweite"
        ],
        "title": "MedienAkteurResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "scheme": "basic",
        "type": "http"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Bundesrepublik API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/content/agenda-ziele": {
      "get": {
        "description": "GET /api/content/agenda-ziele — Spieler-Agenda-Ziele (SMA-501).",
        "operationId": "get_agenda_ziele_api_content_agenda_ziele_get",
        "parameters": [
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "de",
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Agenda Ziele",
        "tags": [
          "content"
        ]
      }
    },
    "/api/content/bundeslaender": {
      "get": {
        "description": "GET /api/content/bundeslaender?locale=de — 16 Bundesländer mit lokalisiertem Namen.",
        "operationId": "get_bundeslaender_api_content_bundeslaender_get",
        "parameters": [
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "de",
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BundeslandResponse"
                  },
                  "title": "Response Get Bundeslaender Api Content Bundeslaender Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Bundeslaender",
        "tags": [
          "content"
        ]
      }
    },
    "/api/content/bundesrat": {
      "get": {
        "operationId": "get_bundesrat_api_content_bundesrat_get",
        "parameters": [
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "de",
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BundesratResponse"
                  },
                  "title": "Response Get Bundesrat Api Content Bundesrat Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Bundesrat",
        "tags": [
          "content"
        ]
      }
    },
    "/api/content/bundle": {
      "get": {
        "description": "Legacy-Bundle: Gesetze kommen aus der DB (Single Source of Truth).",
        "operationId": "bundle_api_content_bundle_get",
        "parameters": [
          {
            "in": "query",
            "name": "scenario_id",
            "required": false,
            "schema": {
              "default": "standard",
              "title": "Scenario Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "de",
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentBundleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bundle",
        "tags": [
          "content"
        ]
      }
    },
...
    "/api/content/chars": {
      "get": {
        "operationId": "get_chars_api_content_chars_get",
        "parameters": [
          {
            "description": "Filtert Chars auf min_complexity <= complexity",
            "in": "query",
            "name": "complexity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 4,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtert Chars auf min_complexity <= complexity",
              "title": "Complexity"
            }
          },
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "de",
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CharResponse"
                  },
                  "title": "Response Get Chars Api Content Chars Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chars",
        "tags": [
          "content"
        ]
      }
    },
    "/api/content/eu-events": {
      "get": {
        "description": "GET /api/content/eu-events?locale=de — EU-Events (Richtlinien, Random, Fix).",
        "operationId": "get_eu_events_api_content_eu_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "de",
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EuEventResponse"
                  },
                  "title": "Response Get Eu Events Api Content Eu Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Eu Events",
        "tags": [
          "content"
        ]
      }
    },
    "/api/content/events": {
      "get": {
        "operationId": "get_events_api_content_events_get",
        "parameters": [
          {
            "description": "Filter: random, char_ultimatum, bundesrat, dynamic, …",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[a-z][a-z0-9_]{0,31}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter: random, char_ultimatum, bundesrat, dynamic, …",
              "title": "Type"
            }
          },
          {
            "description": "Filtert Events auf min_complexity <= complexity",
            "in": "query",
            "name": "complexity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 4,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtert Events auf min_complexity <= complexity",
              "title": "Complexity"
            }
          },
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "de",
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventResponse"
                  },
                  "title": "Response Get Events Api Content Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Events",
        "tags": [
          "content"
        ]
      }
    },
    "/api/content/game": {
      "get": {
        "description": "Narrative content from DB (chars, laws, events, bundesrat). Use ?locale=en for English.",
        "operationId": "game_content_api_content_game_get",
        "parameters": [
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "de",
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Game Content",
        "tags": [
          "content"
        ]
      }
    },
    "/api/content/gesetz-relationen": {
      "get": {
        "description": "GET /api/content/gesetz-relationen?locale=de — SMA-312: requires, excludes, enhances.",
        "operationId": "get_gesetz_relationen_api_content_gesetz_relationen_get",
        "parameters": [
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "de",
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Gesetz Relationen",
        "tags": [
          "content"
        ]
      }
    },
    "/api/content/gesetze": {
      "get": {
        "operationId": "get_gesetze_api_content_gesetze_get",
        "parameters": [
          {
            "description": "Filtert Gesetze auf min_complexity <= complexity",
            "in": "query",
            "name": "complexity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 4,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtert Gesetze auf min_complexity <= complexity",
              "title": "Complexity"
            }
          },
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "de",
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GesetzResponse"
                  },
                  "title": "Response Get Gesetze Api Content Gesetze Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Gesetze",
        "tags": [
          "content"
        ]
      }
    },
    "/api/content/koalitions-ziele": {
      "get": {
        "description": "GET /api/content/koalitions-ziele — Koalitionspartner-Ziele (SMA-501).",
        "operationId": "get_koalitions_ziele_api_content_koalitions_ziele_get",
        "parameters": [
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "de",
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Koalitions Ziele",
        "tags": [
          "content"
        ]
      }
    },
    "/api/content/laws": {
      "get": {
        "deprecated": true,
        "description": "Deprecated: nutze GET /content/gesetze.",
        "operationId": "laws_api_content_laws_get",
        "parameters": [
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "de",
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GesetzResponse"
                  },
                  "title": "Response Laws Api Content Laws Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Laws",
        "tags": [
          "content"
        ]
      }
    },
    "/api/content/medien-akteure": {
      "get": {
        "description": "GET /api/content/medien-akteure?locale=de — Medienakteure mit lokalisiertem Namen.",
        "operationId": "get_medien_akteure_api_content_medien_akteure_get",
        "parameters": [
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "de",
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MedienAkteurResponse"
                  },
                  "title": "Response Get Medien Akteure Api Content Medien Akteure Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Medien Akteure",
        "tags": [
          "content"
        ]
      }
    },
    "/api/content/milieus": {
      "get": {
        "description": "GET /api/content/milieus?locale=de — Milieus mit Ideologie.",
        "operationId": "get_milieus_api_content_milieus_get",
        "parameters": [
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "de",
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MilieuResponse"
                  },
                  "title": "Response Get Milieus Api Content Milieus Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Milieus",
        "tags": [
          "content"
        ]
      }
    },
    "/api/content/politikfelder": {
      "get": {
        "description": "GET /api/content/politikfelder?locale=de — Politikfelder.",
        "operationId": "get_politikfelder_api_content_politikfelder_get",
        "parameters": [
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "de",
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PolitikfeldResponse"
                  },
                  "title": "Response Get Politikfelder Api Content Politikfelder Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Politikfelder",
        "tags": [
          "content"
        ]
      }
    },
...
    "/api/content/verbaende": {
      "get": {
        "description": "GET /api/content/verbaende?locale=de — Verbände mit Ideologie und Tradeoffs.",
        "operationId": "get_verbaende_api_content_verbaende_get",
        "parameters": [
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "de",
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VerbandResponse"
                  },
                  "title": "Response Get Verbaende Api Content Verbaende Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Verbaende",
        "tags": [
          "content"
        ]
      }
    },
...
  }
}